refactor!: extract dependency related function into their own file
