+++ src/index.ts
@@ -4,26 +4,26 @@
  * This module contains utility functions for handling files.
  */
 
+export type { ExtensionDependenciesOptions, ExtensionDependency } from "./dependencies";
+export { getExtensionDependencies } from "./dependencies";
+
 export type {
   CollectOptions,
   ContentTypeResult,
-  ExtensionDependenciesOptions,
-  ExtensionDependency,
   ProcessedFiles,
   VsixFile,
   VsixInMemoryFile,
   VsixLocalFile,
 } from "./files";
+
 export {
   collect,
   getContentTypesForFiles,
-  getExtensionDependencies,
   getExtensionPackageManager,
   isInMemoryFile,
   isLocalFile,
   processFiles,
 } from "./files";
-
 export type {
   ManifestAsset,
   ManifestAssetType,
