+++ test/files.test.ts
@@ -4,7 +4,8 @@ import { join, normalize } from "node:path";
 import { promisify } from "node:util";
 import { assert, describe, expect, it } from "vitest";
 import { fromFileSystem, testdir } from "vitest-testdirs";
-import { collect, getContentTypesForFiles, getExtensionDependencies, type VsixFile } from "../src/files";
+import { getExtensionDependencies } from "../src";
+import { collect, getContentTypesForFiles, type VsixFile } from "../src/files";
 import { readProjectManifest } from "../src/manifest";
 
 const execAsync = promisify(exec);
